requirements.txt
openai
phidata
langchain-community
scikit-learn
sentence-transformers
PyPDF2
pypdf